# swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/auth/signup":{"post":{"operationId":"AuthController_signUp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpDto"}}}},"responses":{"201":{"description":"The user has been successfully created."},"409":{"description":"The email is already exists."}},"tags":["auth"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"Get access token","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenDto"}}}},"400":{"description":"Bad Request"}},"tags":["auth"]}},"/restaurants":{"post":{"operationId":"RestaurantsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRestaurantDto"}}}},"responses":{"201":{"description":""}},"tags":["restaurants"]},"get":{"operationId":"RestaurantsController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["restaurants"]}},"/restaurants/{slug}":{"get":{"operationId":"RestaurantsController_findOne","parameters":[{"name":"slug","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["restaurants"]},"delete":{"operationId":"RestaurantsController_remove","parameters":[{"name":"slug","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}},"tags":["restaurants"]}}},"info":{"title":"Ha Kin - API","description":"Ha Kin API is awersome project for doing with Ha Kin that created from nextjs","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"SignUpDto":{"type":"object","properties":{"name":{"type":"string","example":"John Doe"},"email":{"type":"string","example":"email@gmail.com"},"password":{"type":"string","example":"1234"}},"required":["name","email","password"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string","example":"email@gmail.com"},"password":{"type":"string","example":"1234"}},"required":["email","password"]},"TokenDto":{"type":"object","properties":{"accessToken":{"type":"string","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJuYW1lIjoiYnJvd25pZSBrdW5nIiwiZW1haWwiOiJ1c2VyMUBnbWFpbC5jb20iLCJpYXQiOjE2MjkzOTcyNjN9.ngLiwLRmEstmcFMAoK532cKGTucA8VAglnfrjUa_ABC"}},"required":["accessToken"]},"CreateRestaurantDto":{"type":"object","properties":{"name":{"type":"string","example":"Cozy royale"},"tags":{"example":["burger","lunch","family"],"type":"array","items":{"type":"string"}},"averageCost":{"type":"number","example":300},"images":{"example":["https://images.unsplash.com/photo-1466978913421-dad2ebd01d17?ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=1334&q=80","https://images.unsplash.com/photo-1598514983318-2f64f8f4796c?ixid=MnwxMjA3fDB8MHxzZWFyY2h8NXx8cmVzdGF1cmFudCUyMGZvb2R8ZW58MHx8MHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=800&q=60","https://images.unsplash.com/photo-1457460866886-40ef8d4b42a0?ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=1500&q=80"],"type":"array","items":{"type":"string"}},"phoneNumber":{"type":"string","example":"1-941-871-0498"},"address":{"type":"string","example":"882 Dooley Row Suite 584"},"facilities":{"example":["Home Delivery","WiFi"],"type":"array","items":{"type":"string"}}},"required":["name","tags","averageCost","images","phoneNumber","address","facilities"]}}}}